// src/types/background.ts
export type AutoDeletePeriod =
  | 'never'
  | '30sec'
  | '1min'
  | '1hour'
  | '1day'
  | '7days'
  | '14days'
  | '30days'
  | '180days'
  | '365days'